requirements.txt
asgi-redis==1.4.3
asgiref==3.3.4
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
channels==1.1.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==1.4.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.7
django-adminactions==1.9
django-allauth==0.44.0
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-cleanup==5.2.0
django-crispy-forms==1.11.2
django-etc==1.3.0
django-grappelli==2.15.1
django-grappelli-extras==1.2.3
django-haystack==3.0
django-heroku==0.3.1
django-hitcount==1.3.3
django-import-export==2.5.0
django-indexer==0.3.0
django-js-asset==1.2.2
django-machina==1.1.4
django-mptt==0.12.0
django-suit==0.2.25
django-tawkto==0.3
django-utils-six==2.0
django-widget-tweaks==1.4.8
elasticsearch==7.12.0
elasticsearch-django==7.1.1
elasticsearch-dsl==7.3.0
et-xmlfile==1.1.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
Markdown==3.3.4
markdown2==2.4.0
MarkupPy==1.14
msgpack-python==0.5.6
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==2.10.6
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
stripe==2.57.0
tablib==3.0.0
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.4
webencodings==0.5.1
websocket-client==0.58.0
whitenoise==5.2.0
Whoosh==2.7.4
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
